Rename generationPayloadType to generationTypes
