+++ src/extension/src/types/generationTypes.ts
